+++ src/main/java/uk/gov/companieshouse/exception/DissolutionNotFoundException.java
@@ -1,0 +1,7 @@
+package uk.gov.companieshouse.exception;
+
+public class DissolutionNotFoundException extends RuntimeException {
+    public DissolutionNotFoundException() {
+        super();
+    }
+}
