+++ src/main/java/uk/gov/companieshouse/mapper/DissolutionUserDataMapper.java
@@ -1,7 +1,6 @@
 package uk.gov.companieshouse.mapper;
 
 import org.springframework.stereotype.Service;
-import uk.gov.companieshouse.model.db.dissolution.DissolutionUserData;
 
 @Service
 public class DissolutionUserDataMapper {
