+++ src/test/java/uk/gov/companieshouse/fixtures/DissolutionFixtures.java
@@ -10,9 +10,9 @@ import uk.gov.companieshouse.model.db.dissolution.DissolutionData;
 import uk.gov.companieshouse.model.db.dissolution.DissolutionDirector;
 import uk.gov.companieshouse.model.db.dissolution.DissolutionRejectReason;
 import uk.gov.companieshouse.model.db.dissolution.DissolutionSubmission;
-import uk.gov.companieshouse.model.db.dissolution.DissolutionUserData;
 import uk.gov.companieshouse.model.db.dissolution.DissolutionVerdict;
 import uk.gov.companieshouse.model.db.payment.PaymentInformation;
+import uk.gov.companieshouse.model.domain.DissolutionUserData;
 import uk.gov.companieshouse.model.dto.dissolution.DirectorRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateRequest;
 import uk.gov.companieshouse.model.dto.dissolution.DissolutionCreateResponse;
@@ -129,7 +129,7 @@ public class DissolutionFixtures {
         directorTwo.setName("MERCURE, Fred");
         directorTwo.setEmail("fred@mercure.com");
 
-        return Arrays.asList(director,directorTwo);
+        return Arrays.asList(director, directorTwo);
     }
 
     public static Company generateCompany() {
